+++ tests/testthat/test-cpma.R
@@ -1,6 +1,7 @@
 # default behavior
 p_2 <- list(0.05, 0.05)
 p_10 <- as.list(rep(0.05, 10))
+p_ns <- as.list(rep(5E-8, 10))
 
 test_that("cpma works for two p-values", {
   expect_equal(cpma(p_2)$p, 0.05798245, tolerance = 1E-5)
@@ -12,6 +13,11 @@ test_that("cpma works for ten p-values", {
   expect_equal(cpma(p_10)$p, 2.243116e-05, tolerance = 1E-5)
   expect_equal(cpma(p_10)$n, 0)
   expect_equal(cpma(p_10)$p_exp, -log10(2.243116e-05), tolerance = 1E-5)
+})
+
+test_that("cpma works the same as non-subset cpma", {
+  expect_equal(cpma(p_2)$p, do.call("p_cpma", p_2))
+  expect_equal(cpma(p_ns)$p, do.call("p_cpma", p_ns))
 })
 
 # handling exceptions
@@ -44,8 +50,10 @@ test_that("cpma can deal with illogical values", {
 # handling extreme values
 p_0 <- list()
 p_1 <- list(0.05)
+p_10k <- as.list(rep(0.5, 10000))
 
 test_that("cpma can deal with extreme values", {
   expect_error(cpma(p_0))
   expect_error(cpma(p_1))
+  expect_equal(cpma(p_10k)$p, 1.827719e-261, tolerance = 1E-5)
 })