+++ tests/testthat/test-fct.R
@@ -1,6 +1,7 @@
 # default behavior
 p_2 <- list(0.05, 0.05)
 p_10 <- as.list(rep(0.05, 10))
+p_ns <- as.list(rep(5E-8, 10))
 
 test_that("fct works for two p-values", {
   expect_equal(fct(p_2)$p, 0.01747866, tolerance = 1E-5)
@@ -12,6 +13,11 @@ test_that("fct works for ten p-values", {
   expect_equal(fct(p_10)$p, 7.341634e-06, tolerance = 1E-5)
   expect_equal(fct(p_10)$n, 0)
   expect_equal(fct(p_10)$p_exp, -log10(7.341634e-06), tolerance = 1E-5)
+})
+
+test_that("fct works the same as non-subset fct", {
+  expect_equal(fct(p_2)$p, do.call("p_fct", p_2))
+  expect_equal(fct(p_ns)$p, do.call("p_fct", p_ns))
 })
 
 # handling exceptions
@@ -40,8 +46,10 @@ test_that("fct can deal with illogical values", {
 # handling extreme values
 p_0 <- list()
 p_1 <- list(0.05)
+p_10k <- as.list(rep(0.3, 10000))
 
 test_that("fct can deal with extreme values", {
   expect_error(fct(p_0))
   expect_error(fct(p_1))
+  expect_equal(fct(p_10k)$p, 4.113606e-82, tolerance = 1E-5)
 })